rename useNewActivity to useActivityForm
